Support multiple modules in one package directory.
If you call terraform_module multiple times in one Bazel buildfile, you would previously see an error because each module was creating its own :terraform target.

With this change, the Terraform executable target will be prefixed with the name of the terraform_module, unless that name matches the name of the package directory.